bug 1743510: capture minidump-stackwalk output and surface it
This reworks `execute_process` so that the `MinidumpStackwalkRule` can
surface stderr as `mdsw_stderr`. It adds that to the Debug tab so it's
visible.

The combination of the exit code plus the last lines of stderr will make
debugging minidump-stackwalk easier plus it tells us a lot more about
what's bad about specific minidumps.