bug 828452: add signature generation api
This adds an API endpoint for signature generation: /api/CrashSignature/

It's an HTTP POST API where the payload is one or more jobs in the form
that the signature generation library takes.

This adds django-rest-framework and reimplements the CrashVerify API to
use it. It's also used to implement the new CrashSignature API.

This adjusts the API documentation code to support SocorroAPIView
classes (which are django-rest-framework classes), though the support is
not great. I'll work on improving this over time. It's probably the
case we'll wipe out model_wrapper and turn everything into a
django-rest-framework API class.