bug 1745663: fix error codes in crash signatures
When minidump-stackwalk encounters an error (empty minidump, no threads
stream, etc), the MinidumpStackwalkerRule should stomp on the
mdsw_status_string with that error code and then signature generation
can grab it to use in the signature.

This updates rust-minidump to ebb603f5:

ebb603f Start top-level errors with an error name.

Note that this changes some crash signatures.

With breakpad stackwalker, we'd have signatures like:

* EMPTY: no crashing thread identified; ERROR_NO_MINIDUMP_HEADER
* EMPTY: no crashing thread identified; ERROR_NO_THREAD_LIST

With rust-minidump minidump-stackwalker and these changes, we'll now
have:

* EMPTY: no crashing thread identified; EmptyMinidump
* EMPTY: no crashing thread identified; MissingThreadList