bug 1745663: fix KeyError caused by CrashingThreadRule
I made a change to handle 0-byte minidumps and skip running
minidump-stackwalk. That's fine, but I thought it made sense to "fake"
the stackwalker_data output. That turns out to cause problems with other
processor rules that have expectations for what is and isn't in the
minidump.

This scales that back a bit so that json_dump isn't set to {} which
caused the events that lead to a KeyError.