bug 1746872: overhaul crash reporting spec
This cleans up the crash reporting spec and reorganizes it so
(theoretically) it's easier to follow for someone who's writing a crash
reporting client.

This adds some additional HTTP response codes and what the client should
do.

This moves the notes about differences with the Breakpad upload tool to
a section at the end.