bug 1749561: fix supersearch refine links
Previously, if you do a supersearch and refined the search by clicking
on one of the field values in the results or facet results, it would add
that as an "is" query to the search. That's fine, except when there's a
filter for that field. In that case, the "is" query is OR'd with the
existing filter and I can't think of a single case where that's what you
want.

Now the refine links replace any existing filters for the field with the
"is" query.

Further, it fixes product, version, and product_type refinements to use
a term query (no prefix) instead of an "is" query ("=" prefix) so that
it works correctly with the supersearch form and site navigation.

Then I moved the code to macros because the template is really messy and
hard to read and we have a lot of complex-looking repeated logic
everywhere.