bug 1748340: handle empty list product_name situation
If the passed in product is an empty list (which I think is possible
because of the way the parameters are determined), then the correlations
page would die because there's no index 0 in an empty list.

This accounts for the empty list possibility by using the default
product if the determined product name is falsey (empty string, empty
list, not set, ...).