bug 1749561: fix supersearch form and refining search
This fixes the javascript parts of the supersearch form when you go to
refine the search.

It correctly removes the "exists" filters if they exist for the field.
It no longer messes up adding a filter for mac_crash_info or other
fields where the values get coerced into other things.

It still adds filters for product, product_type, platform, and version
rather than adding those to the simple form. I thought about trying to
figure that out and then decided it's not _critical_, so I'd leave it.