Fix intermittently failing test
The integration test with sentry-sdk and fakesentry fails occasionally.
I think what's going on is that the test didn't give enough time for
sentry-sdk to send the error to fakesentry before checking to see if the
error was there. This adds a time.sleep() to pause. I think this will
fix it. If not, we should look further into the timing of things by
setting debugmode in sentry-sdk.