bug 1751799: capture total ingestion time
This adds some code to capture the total time it takes to ingest a crash
report from collection (submitted_timestamp) to the end of processing.
This coupled with the queue size helps us understand the backlog and the
effect it has on how up-to-date the data we're looking at is.