bug 1753550: recover after cache.set error
Apparently we have some data we want to cache that's larger than the
memcached max size and that's causing cache.set() to error out.

This wraps that in a try/except block and emits a cache_set_error metric
so the code "recovers", but also lets us monitor the situation in case
we want to do something different when we learn more.