bug 1743704: remove jit crash classifier
The JitCrashClassifierRule runs _after_ signature generation, but
signature generation has a SignatureJitRule that prefixes "jit" to
signatures that have "jit_cateogry" which comes from
"classification.jit.category" which is set by JitCrashClassifierRule.
That's circular and thus signatures will never have "jit" prefix unless
they were reprocessed!

jit-crash-classifier comes from the old Breakpad-based
minidump-stackwalk which we stopped using in December
2021 in favor of rust-minidump. rust-minidump doesn't have a
jit-crash-classifier equivalent (yet). In a discussion about this
situation, it was said that the jit-crash-classifier often has false
positives, so it's not useful.

We're not removing the fields from telemetry_socorro_crash.json--the
processor will omit them in the processed crash data files it generates.
I added some notes to this effect to the schema file.