bug 1750759: fix path tag value and sort tags
This sorts the tags alphabetically because influxdb docs say that's a
good thing.

This also removes the / from the tag path values. I think that's the
cause of the issue I'm hitting in stage where tag path values are all
showing as "undefined" in Grafana. This is a bit of a guess, though,
since I can't find good documentation on what makes a valid tag value.