bug 1753521: fix accessibility_client and accessibility_in_proc_client
This changes accessibility_client and accessibility_in_proc_client to be
copied from the raw crash to the processed crash and changes supersearch
to look at the processed crash value.

Once the old crash data expires from the system, we can change this to
only look at the processed crash.