bug 1753521: fix android_cpu_abi and android_cpu_abi2
This changes android_cpu_abi and android_cpu_abi2 to be copied from the
raw crash to the processed crash and changes supersearch to look at the
processed crash value.

Once the old crash data expires from the system, we can change this to
only look at the processed crash.