bug 1753521: fix available_physical_memory and available_virtual_memory
This changes available_physical_memory and available_virtual_memory to
be copied from the raw crash to the processed crash and changes
supersearch to look at the processed crash value.

Once the old crash data expires from the system, we can change this to
only look at the processed crash.