bug 1753521: fix content_sandbox_capabilities and content_sandbox_level
This changes content_sandbox_capabilities and content_sandbox_level to
be copied from the raw crash to the processed crash and changes
supersearch to look at the processed crash value.

Once the old crash data expires from the system, we can change this to
only look at the processed crash.