bug 1753521: fix startup_time
This changes startup_time to be copied from the raw crash to the
processed crash and changes supersearch to look at the processed crash
value.

This redoes the DatesAndTimesRule because startup_time comes from the
processed_crash and it gets validated in the copyover.

Once the old crash data expires from the system, we can change this to
only look at the processed crash.