bug 1753521: fix total_page_file, total_physical_memory, and total_virtual_memory
This changes total_page_file, total_physical_memory, and
total_virtual_memory to be copied from the raw crash to the processed
crash and changes supersearch to look at the processed crash value.

Once the old crash data expires from the system, we can change this to
only look at the processed crash.