bug 1755528: redo how we handle flags and boolean fields
This redoes how we handle flags and boolean fields that we currently
support and how we're going to handle them going forward.

We will validate and normalize values using these rules:

1. if the annotation doesn't exist, nothing gets added to the processed
   crash
2. if the annotation value is truthy, the processed crash gets a True
3. if the annotation value is falsey, the processed crash gets a False

Then we'll index these values as booleans with minimal validation
logic--that's handled in the processor now.

This fixes WindowsErrorReporting annotation accordingly.