bug 1756223: update docker image and remote docker version
This switches from ci-docker-bases to cimg/python. In doing that, we
bumped into an issue where suddenly we were using docker-compose v2
which doesn't support configman environment variables.

This forces circleci to use docker-compose 1.29.2 which is available
from PyPI.

This stops doing work in /socorro which the logged in user doesn't have
access to.

This also updates the remote docker version.