bug 1758701: switch sentry-wrap.py to click
This switches sentry-wrap.py to use click instead of argparse. We've got
click in the requirements and it's easier to use for subcommands, so we
should switch.

This also improves the error handling situations so it's clearer what
happened when something happened.

This also adds timings to the success/fail status output.