bug 1759080: attempt to reduce intermittent ci fail
The TestIntegrationSuperSearchFields.test_test_mapping test fails
intermittently in CI. It feels like a race condition between creating
the new index, ES figuring its stuff out, and then running a search on
the cluster.

This adds some code to make sure the cluster is happy before moving on
to the search. It also wraps the search code in a loop to re-attempt a
few times before ultimately failing.

Maybe that'll help.