bug 1759591: rewrite ModuleURLRewriteRule
The ModuleURLRewriteRule pulls the url for the symbols file for a module
from the minidump-stackwalk output and rewrites it to be "more correct"
in most cases. This fixed problems with how the old stackwalker found
symbol files, cached the file locally on disk, but never kept track of
where it found the symbol file. Because of that, some modules ended up
with the wrong symbol file url. Rewriting that url made most things
mostly better.

The new minidump-stackwalk nicely adds the source url for the symbol
file to the end of the symbol file when it caches it on disk. This
alleviates the problems the old one had. However, there are some cases
where the symbol url is "wrong" or we don't want to reveal the symbol
file source. Plus the symbol url has querystring parameters added on
that help us track missing symbols in the symbols server that we don't
need in crash ingestion.

This rewrites the ModuleURLRewriteRule [1] to redact some urls and
remove the querystring for others.

[1] I decided not to call it RewrittenModuleURLRewriteRule2022. Good
idea? Missed opportunity? Let me know in the comments!