bug 1754035: add test to validate schemas are valid schemas
We had a validate_java_exception.py script which validated that one
file, but it wasn't executed in CI and it only validated one of the two
schemas we have.

This removes that script and adds a test that validates all the schemas
in that directory.

This also improves the schema documentation.