bug 1755528: fix startup_crash
Pull startup_crash from the processed crash instead of the raw crash.

One caveat of this is that I left the "null_value is False" thing in
which makes this different from other flags. We used startup_crash in
the signature report and top crashers reports in some interesting ways.
We might be able to change this, but it requires more analysis than I
have time for now.