bug 1755095: fix TelemetryBotoS3CrashStorage to not use super search fields
This redoes TelemetryBotoS3CrashStorage to build the crash document it
sends to telemetry ingestion solely from the processed crash.

This no longer uses super search fields and reduces some of the tangling
between crash storage systems which is creating a lot of problems.

The processed crash contains data that has been normalized and
validated, so it's a better source to use than the raw crash. This
handles a few historical naming issues, but otherwise it's pretty
straight-forward and driven entirely by the telemetry_socorro_crash.json
schema.