bug 1762000: remove vestigial Winsock_LSP code
In 2019, we removed the processor rule that copied Winsock_LSP from the
raw crash to the processed crash. However, we didn't delete the
super search field entry. This removes that.

This removes this field from the tests--it's not doing anything.

This changes the description in the telemetry_socorro_crash.json schema
to note that this field is empty and deprecated.

This removes the field from the allow list for the processed crash API
since the field isn't in the processed crash.