bug 1762271: rewrite json schema reducer
This drops json-schema-reducer library for one we wrote that's more
correct in how it traverses the schema, does some light type checking,
and allows for an include predicate which we'll use later for redacting
protected data from documents.

This is based on the json schema draft-04 spec which is what our schemas
use.

It doesn't implement some of the data restrictions specified in
draft-04 (for example minItems, maxItems, etc) because we're not
currently using them.