bug 1767311: add support for SubmittedFrom
This adds support for the new SubmittedFrom annotation and also adjusts
the existing support for the SubmittedFromInfobar annotation which is
being phased out.

I added submitted_from to super search fields.

I added SubmittedFrom, SubmittedFromInfobar, and Throttleable to the
Debug tab to make it easier to debug crash reporting and collection
problems.