bug 1394867: make CoUnmarshalInterfaceResult public
This makes CoUnmarshalInterfaceResult public and it also adds it to the
processed crash.

The value is in hexidecimal. I'm copying it over as a string figuring
it's more useful that way as opposed to converting it to decimal. We can
change it if the theory turns out to be false.