bug 1764395: implement processed crash schema
This implements a processed crash schema that covers:

* all the fields that get added to the processed crash during processing
* the schema of the Breadcrumbs crash annotation
* the schema of the JavaException crash annotation
* the schema of the rust-minidump stackwalker output

I implemented a processed crash schema validator that validates the
schme as valid jsonschema and also runs the schema over a directory of
processed crash files. That makes it more likely the schema covers all
the variations of our crash data.

While working on this, I added missing descriptions in super search
fields, added some comments to things that should get fixed at some
point, fixed the yaml item in .editorconfig, enhanced the reducer code
to make it easier to log all the keys in a document,