bug 1764395: add traverse_schema and redo schema tests
This adds traverse_schema which does an preorder traversal of the schema
document executing a specified function. This adds tests for that as
well. Generally, I implemented the traversal bits I needed to support
for our schemas.

This redoes the schema tests and now verifies that the socorro metadata
section is correct.