bug 1764395: fix validate_processed_crash to calculate missing bits
This removes the big commented out section of validate_processed_crash
and adds the bits required to figure out which keys were in the
document, but not the schema and which were in the schema, but not the
document.

This has one caveat: patternProperties. This still doesn't handle that
scenario correctly.