bug 1764395: implement permissions_predicate and annotate permissions
This implements a permissions_predicate that reduces a document based on
the required permissions set on fields and the permissions the user has.

This adds tests for the permissions_predicate.

This annotates the processed crash schema with permissions derived from
the API_ALLOWLIST from the ProcessedCrash API code coupled with the
permissions from super_search_fields.py because both were involved in
non-overlapping ways (which is one of the reasons we're moving
permissions to a single central schema).

This adds another validation step to the processed crash schema
validation where we verify that fields all fields have permissions set,
fields that are public don't have non-public parent fields because
that's definitely a mistake, and that the set of public fields is equal
to the expected set of public fields so as to reduce mistakes.

This adds a few more fields to the protected crash schema for things
that should have been in the processed crash and derived from crash
annotations but somehow got lost in the shuffle over the last month
(which is another reason we're moving processed crash management to a
schema instead of strewn about the codebase as a series of unfortunate
side-effects).

This adds a few more fields to the socorro metadata that aren't used
now, but will be at some point.