bug 1764395: rewrite schemas and reducers and traversers
This is a semi-big rewrite of a lot of stuff.

First, I changed the plan. Instead of defining a processed crash schema
(jsonschema), we now define a socorro data schema (jsonschema) and then
define the processed crash schema (socorrodata). The socorro data schema
codifies the bits we were calling "socorro metadata" before as part of
the schema. It enforces some of the things we were enforcing in
validation tests. It is structurally similar to jsonschema (types are
the same, object and array traversal is a subset) without some of the
complexity we don't need (items as records vs. items as lists). I
copyied some bits from the glean-parser metrics schema so it's as close
to that as we can make it. Overall, I think this is a big improvement.

However, it does mean we need two different sets of traversing things:
one for jsonschema schemas (telemetry crash schema) and one for
socorro-data schemas (processed crash).

We can share some things. I made "$ref" in socorro-data the same-ish as
jsonschema. The basic types are the same. socorro-data uses properties,
pattern_properties, and items in the same way as jsonschema (roughly).

I wrote a resolve_references function that takes a schema, resolves all
the references, and drops the "definitions" section.

For jsonschema things, I wrote a JsonSchemaReducer which we use in
TelemetryBotoS3CrashStorage. It reduces a document according to the
shape of the specified schema. It no longer does other transformations
or predicates or whatever, so it's simpler. It also doesn't support the
full jsonschema spec--only the bits I needed.

For socorro-data things, I wrote a SocorroDataReducer that reduces a
document according to the shape of a specified schema. It doesn't do any
other transformations. I wrote a transform_socorro_data_schema function
that takes a schema and a transform function and returns a new schema.
We'll use that for handling permissions.

I updated TelemetryBotoS3CrashStorage to create a public-only derived
schema from the processed crash schema. It reduces the processed crash
document with that so we're gating what's public and what's not from the
processed crash schema. It then reduces the public-only processed crash
data using the telemetry schema giving us a document in the right shape
for telemetry ingestion.

I rewrote the ProcessedCrash API to create a series of
permission-variations derived schemas from the processed crash schema.
It uses the appropriate one for reducing the processed crash to the data
the user is permitted to see. No more API_ALLOWLIST nonsense.

I updated the processed crash schema to be valid socorro-data. I added
permission annotations to it with permissions derived from the
API_ALLOWLIST from the ProcessedCrash API code coupled with the
permissions from super_search_fields.py because both were involved in
non-overlapping ways (which is one of the reasons we're moving
permissions to a single central schema

I also added a few fields to the processed crash schema that I think got
lost in the shuffle.

This forces dom_fission_enabled and dom_ipc_enabled to get converted
from a boolean to a string because that's what Telemetry ingestion is
expecting.

I fixed the processor to add isoformat strings to the processed crash
for client_crash_date, started_datetime, completed_datetime, and
submitted_timestamp. They get saved to crash storage as isoformat
strings and when the processed crash is loaded, they stay as isoformat
strings. Given that, it's much easier to _always_ have them be isoformat
strings.

The memory report can contain urls of pages, so it should never be
public. This adds a comment to the processed crash schema and adjusts
the description in the telemetry socorro crash schema.

cpu_count and stackwalk_version were indexed into Elasticsearch from
processed_crash.json_dump. The cpu_count is in the processed crash
already, so we can/should pull the value from there like the other cpu_
fields. The stackwalk_version field should be up a level as well and
match the mdsw_ fields.

This moves stackwalk_version up a level and adds a note in the
report_index.html code to drop the backwards-compatible code in 12/2022.

This moves cpu_count up a level and adds a migration and a note to
super_search_fields.py to drop the backwards-compatible code in 12/2022.