bug 1764395: rewrite ProcessedCrash API to use reducer
This redoes the ProcessedCrash API to pull the processed crash data and
then reduce it using the processed crash schema and the user's
permissions. No more API_ALLOWLIST nonsense for the ProcessedCrash.

After doing that, I had to fix a bunch of tests which were using test
data that didn't line up with the schema (or reality).