bug 1764395: fix cpu_count and stackwalk_version
These two fields were indexed into Elasticsearch from
processed_crash.json_dump. The cpu_count is in the processed crash
already, so we can/should pull the value from there like the other cpu_
fields. The stackwalk_version field should be up a level as well and
match the mdsw_ fields.

This moves stackwalk_version up a level and adds a note in the
report_index.html code to drop the backwards-compatible code in 12/2022.

This moves cpu_count up a level and adds a migration and a note to
super_search_fields.py to drop the backwards-compatible code in 12/2022.