bug 1764395: split libjson and clean up jsonschema code
"libjson" was a serious misnomer since the module didn't have anything
for manipulating json in it.

This splits libjson into libjsonschema and libsocorrodataschema. The
former holds jsonschema things to manipulate the telemetry crash data
schema which is a jsonschema schema. The latter holds scorrodataschema
things for the processed crash.

I also removed the "required" validation because we don't need it
jsonschema.