bug 1764395: remove java_exception_raw
Previously, we couldn't handle permissions being different in fields in
nested structures so we did this thing where we'd have a "raw" and
"redacted" copies of data structures in the processed crash.

We now have the ability to set the permissions for fields in nested
structures and prevent them from showing up for people without those
permissions.

Given that, this removes java_exception_raw and all the code to support
that.