bug 1764395: return all data if user has view_pii for ProcessedCrash API
This changes the ProcessedCrash API to return all the data without
reducing it if the user has view_pii permissions. This matches what is
currently in prod and makes it easier to see the final processed crash
data including stackwalker output even if that's not covered by the
schema.

We still have the scaffolding to do more granular permissions for if we
do something like re-add view_exploitability or something like that.