bug 17885331: fix Accessibility processing
The Accessibility field is in a crash report and set to "Active" by the
accessibility service when it is active. Otherwise it's not in the crash
report at all. The processed crash schema expects a boolean. Super
Search expects a boolean. The telemetry schema expects a string and the
value that Socorro sent to telemetry was "Active".

This reconciles all this down to setting the value in the processed
crash to True if the annotation exists and the value is "Active" and
False otherwise. It converts that to a "1" for Telemetry.