bug 1788971: fix jsonschema deprecation warning
jsonschema was kicking up this deprecation warning:

```
DeprecationWarning: The metaschema specified by $schema was not found.
Using the latest draft to validate, but this will raise an error in the
future.
```

This fixes that by passing in the correct validator. Further, since I
think I want to be validating test data because a lot of it is wrong, I
created a function for it that we can use everywhere.