bug 1604845: remove entropy from raw crash keys
We added an entropy part to the keys for raw crashes a long time ago so
as to increase performance when saving and loading things in rapid
succession to S3 buckets. Those times are long gone and that's not a
helpful optimization.

This removes that part from the path so as to simplify things like
directory listings for raw crashes on a certain date.

This also fixes permadelete_crash_data, remove_field, and
fetch_crash_data to use build_keys to build keys rather than having the
key structure scattered across the codebase.