bug 1604845: switch raw crash key version to v1
We're effectively returning to the original v1 from ages ago. Version
changing shenanigans are irritating, but in this specific case, I don't
think anything else relies on the pathing except antenna and socorro.
Given that I control both and I'm changing it anyhow, it seems easier in
general to return to v1 which matches the other paths than to keep v2 or
go to v3.