bug 1710725: stop using DotDict for raw and processed crash data
This stops using DotDict for raw and processed crash data. When data is
loaded from crash storage, it'll now be in standard Python data
structures (dict, list) and needs to be accessed as such.

We were doing this already for the most part, so this is a minor change
at this point.