bug 1710725: replace processor_meta_data with Status instance
The Socorro processor pipeline keeps notes related to processing as
crash reports wind their way through the rules. At the end, the
processor notes are added to the crash report. In this way, we have some
visibility into processing.

Previously, processor_meta_data was a DotDict with a handful of fields.
In the previous commit, we changed it so it was a dict. In this commit,
we remove all the unused fields and then change it to a data class with
a more convenient way of adding notes. Then we update all the rules and
code that uses it.