bug 1789705: fix table column widths across the site
I threw together some width classes in an earlier commit. This tweaks
the widths of the stacks table to fit the data better. This also
improves a bunch of other tables to make them more stable layout-wise
and remove pixel widths.

I also adjusted borders for table headers. There was a bunch of css code
in random places for that. I reduced that some.