bug 1767566: fix clickability of tabs in report view
The tabs are implemented with jquery ui tab component and it's
structured such that the "tab" is a li with an a child that responds to
click events. But since the tab has aria-labeledby that tells screen
readers to ignore the content, the a isn't clickable.

This adds James' suggested fix which adds a listener for click to the
tab which then clicks the a tag.