bug 1759065: stop building stackwalker every build
We've moved building the stackwalker to a separate repository that does
"releases" which include a .tar.gz file. The file is like 5mb.

Then we changed the Socorro build so that it downloads the file rather
than building stackwalker _again_. This will improve build times when
the docker image cache isn't available or doesn't have anything useful.