bug 1687987: add code to migrate raw crash structure
This adds some code to migrate the raw crash structure from version 1 to
version 2. The new structure puts most of the collector-generated data
into a "metadata" section. This reduces the possibility of key
conflicts, makes it easier to discern annotations from
collecter-generated things, and makes it easier to write a schema and
apply permissions.