bug 1787931: switch to using get_schema() and fix BreadcrumbsRule
This switches everything that's accessing the PROCESSED_CRASH_SCHEMA to
getting it from socorro.lib.libsocorrodataschema..get_schema which gets
the schema and resolves references.

The BreadcrumbsRule needed to access the part of the schema that changes
when references are resolved and that created some issues and required
some deepcopy stuff. This adjusts the rule to use the structure from the
schema after references are resolved.