bug 1787931: add default_permissions to schema
This moves the "default permissions" to the schema itself rather than
have it dribbled around the codebase. If we ever need to add another
permission, now we can do it in this one file.

When doing this, I found we had missed defining a permission for inlines
properties. This fixes that, too.