bug 1797295: fix create-a-bug links
The create-a-bug links are generated with a Jinja2 function which uses a
Jinja2 template to generate the comment. Jinja2 templates are great for
HTML where consecutive whitespace is no big thing, but they're awful for
text output especially complex templates with if/then/else and for
loops.

Over the years, we've spent a lot of time trying to get the
bug_comment.txt template to work for the variations it needs to support.

This adds a "look, squash all cases of 2 or more line endings to just 2"
re.sub() which makes editing the template a lot less nightmarish.