bug 1797742: show unloaded modules in modules tab
This adds the unloaded modules to the modules tab in the report view.

One thing to note is that unloaded modules show up multiple times in the
list, but have different base/end addresses. Since we're not displaying
the base/end addresses in the list, it seems like there's a bug causing
them to be repeated. We could not repeat them, but it seems interesting
to know that a module shows up three times. We could show the base/end
addresses, but I was on the fence about that. We can think about that
another day.