bug 1796389: add "stackoverflow" to signature
This adds a new tag "stackoverflow" to crash signatures when the reason
(json_dump.crash_info.type) is "EXCEPTION_STACK_OVERFLOW". Because this
is only one of a myriad of indicators and we want to expand support over
time, I structured the rule like the OOMSignature rule which also has
multiple indicators.