bug 1795762: show current signature if signature generation has changed
If signature generation has changed since when a crash report was last
processed, then the report view will now show a note with the new
signature.

In order to do this, I needed to move some things around, so I took some
time to make the top of the report view more consistent with other parts
of the site and reduce some of the complexity in design.