bug 1784558: fix collector_metadata issues
This adds collector_metadata to the processed crash schema and fixes the
descriptions in the raw crash schema to be more correct looking.

This fixes super search fields so that collector_notes is a field, but
it pulls from a nested structure. That's something we've only done with
migrations, so now we're going to do it generally. I had to add another
clause to the test regarding the source key, but we might want to nix
that test altogether.