bug 1792255: add telemetry_environment to processed crash
This adds the telemetry_environment bits to the processed crash
schema. I cribbed a lot of the types and descriptions from:

https://github.com/mozilla-services/mozilla-pipeline-schemas/blob/d7c892dbfe46c37c44d897b8edb8b0d9f5273fec/schemas/telemetry/event/event.4.schema.json

A lot of fields don't have descriptions. That's unfortunate. We can fill
those in as we go along, I guess.

I didn't include things that weren't in the crash reports I spot
checked. Whatever isn't in the schema will be treated as protected data,
so we can add it as we need it without risk of a data leak.

I also fixed the report view and template to use the version in the
processed crash. It's normalized and validated now, so it's less likely
for weird data to cause an error in the report view.