bug 1792255: fix validate_processed_crash for pattern_property parts
This fixes an issue in match_key where schema keys with pattern_property
regex parts didn't match document keys because of the way we represented
the path as a .-delimited string of parts. This is a very ugly fix, but
the "better" fix involves rewriting all the code around paths during
traversals to keep a list of parts rather than a .-delimited string.