bug 1808593: rewrite c/s/models.py get_api_allowlist
This rewrites the crashstats/supersearch/models.py get_api_allowlist()
so it makes more sense with the current architecture. Many years ago, we
moved super search fields from being defined in the database (where they
were mutable) to being defined in a Python module (where they aren't).

get_api_allowlist() was written when the fields were mutable. Now that
they aren't, we can take a lot of the dynamicism out of it.