bug 1808594: remove measure_fetches
This removes the code for measuring fetch timings and cache hits/misses
as well as the admin page to view the data. While this information is
interesting, we weren't doing anything with it and we're sending metrics
timings to Grafana, so this is a little redundant.

Thus we're removing it.