bug 1808595: stop caching RawCrash and ProcessedCrash data
This changes the RawCrash and ProcessedCrash models to stop caching
results. This updates the report view and drops the "refresh=cache"
bits, too.

All caching is doing here is reducing the number of times we fetch data
from S3. For the report view, HTTP GET requests are sub-1s, don't happen
often, and don't tend to repeat. If we need to cache things, we should
cache the output of rendering the report view template which would save
us two S3 fetches and the work to render the template. We can look into
that if we encounter a need to make this faster.