bug 1803558: add nickname for pattern_properties
In order to document a pattern_properties field, we really need it to
have a name so we can refer to it. pattern_properties field names are
regular expressions and that's not useful for referring. This adds an
informal nickname we can use because "crumb" is clearer than "^.+$".

While doing this, I fixed a bug in the socorro-data-1-0-0 schema in the
pattern_properties definition. I also cleaned up some goofy stuff
involving unhelpful uses of allOf and anyOf.