bug 1807956: truncate stacks in processor
This adds a rule to the processor to truncate stacks in the stackwalker
output. Stacks can sometimes be egregiously large and stacks that large
often contain repeated sequences of frames and the repetition isn't
helpful.

This picks a MAX_FRAMES of 500 figuring that's enough frames to get a
sense of things. If you need more frames, you should download the
minidump, run the stackwalker, and do your analysis on the raw
stackwalker output.