bug 1807956: remove truncation in the webapp
We previously implemented frames truncation in the webapp where if the
stack was > MAX_FRAMES (in this case, 100 frames), it would truncate it.

Now that we're truncating stacks in the processor (where MAX_FRAMES is
500), it makes sense to remove stack truncation in the webapp.

For all the crash reports up to now that have egregious stacks, we can
reprocess them and that'll truncate the stacks and then we'll be fine.