bug 1809451: improve symbolssupplier
This adjusts the run script to create the symbols directory first. This
adds the symbols directory to .gitignore (don't want to check in
contents) and .dockerignore (don't want contents to be part of docker
image). This  adds symbolsserver to the services that run with `make
run`. This adds it as the second item in the symbols_url for the
processor.