bug 1767282: rework temp directory use in processor
The processor uses a temporary directory to save crash report data in
order to process it. Previously there were multiple parts of the
processor saving data in different places--some of which weren't safe
from thread contention.

This fixes that by pulling ownership of the temporary directory
(creation, deletion) into the processor app and then passing that
directory around to the pipeline and rules. The processor app now
creates a new temporary directory in process_crash that's unique to that
processing run. This should fix orphaned files and thread contention
issues (and future possible process contention issues).