bug 1767282: cleanup stackwalker error handling
This undoes a change I just made where I added the last bit of stderr to
the logs. Instead, this cleans up the error handling so that SIGKILL and
SIGABRT are logged more helpfully and non-0 errors are handled in one
place rather than a few places. This also makes sure that arbitrary
stderr data doesn't make it into the processor notes.