bug 1529342: merge FetchTrasnformSaveApp with ProcessorApp
The ProcessorApp was the only subclass of FetchTransformSaveApp and
there was a bunch of extra code between the two to create an API that
allowed us to build the ProcessorApp. Given that ProcessorApp is the
only subclass, it makes sense to merge the two and tidy up the code that
was around the API barrier.

While doing this, I also tidied up the crash processing code so that the
resulting "process_crash" method was minimal with all the context set up
done by the caller.