bug 1815235: stringify cond names in SignatureShutdownTimeout rule
We hit a case where the condition name was a float. That should never
happen and we should catch that in processing in a
validation/normalization rule so that signature generation doesn't have
to deal with weird data.

However, in the interests of getting a fix out there now, I'm just going
to stringify the condition names.