bug 1529342: remove configman from verifyprocessed and archivescraper
This starts removing configman from configman_utils using things in the
webapp.

This removes configman bits from archivescraper--I think it was a
comment.

This reworks esclean.

This removes configman from verifyprocessed. It also moves some code out
of verifyprocessed and into the boto crashstorage code. This will make
it easier down the road when we switch to GCP because the storage code
is in one place.

This removes configman from updatemissing and updates updatemissing to
use the updated bits in verifyprocessed.