bug 1529342: reimplment timing/incr metrics for saving processed crashes
This reimplements how the timing/incr metrics are emitted when saving
processed crashes. Previously, they were generated by a MetricsCounter
faux crash storage class (the incr) plus benchmarking classes that
wrapped all the crash storage classes (the timings).

This ditches all that and adds the incr and timing code to the processor
app process_crash method. In doing this, I made the keys consistent with
one another and based on the crash destination name (i.e. "s3", "es",
"telemetry). For example, keys for elasticsearch crashstorage started
with either "processor.es" or "processor.elasticsearch". Now they start
with "processor.es".

Old keys:

* incr processor.elasticsearch.raw_crash_size
* incr processor.elasticsearch.processed_crash_size
* incr processor.elasticsearch.crash_document_size
* timing processor.elasticsearch.index
* timing processor.s3.BotoS3CrashStorage.save_processed_crash
* timing processor.es.ESCrashStorage.save_processed_crash
* timing processor.telemetry.TelemetryBotoS3CrashStorage.save_processed_crash

New keys:

* incr processor.es.raw_crash_size
* incr processor.es.processed_crash_size
* incr processor.es.crash_document_size
* timing processor.es.index
* timing processor.s3.save_processed_crash
* timing processor.es.save_processed_crash
* timing processor.telemetry.save_processed_crash

When this is deployed, we'll need to update the dashboards.